Rename tmp name for ESLint Unicorn.
